Convert system tests code to ESM
See https://jestjs.io/docs/ecmascript-modules
