+++ system-tests/001/fs.test.mjs
@@ -1,7 +1,6 @@
-'use strict';
+import stylelint from '../../lib/index.js';
 
-const stylelint = require('../../lib');
-const { caseConfigFile, caseFiles, prepForSnapshot } = require('../systemTestUtils');
+import { caseConfigFile, caseFiles, prepForSnapshot } from '../systemTestUtils.mjs';
 
 const CASE_NUMBER = '001';
