+++ system-tests/002/fs.test.mjs
@@ -1,9 +1,8 @@
-'use strict';
+import postcssScss from 'postcss-scss';
 
-const postcssScss = require('postcss-scss');
+import stylelint from '../../lib/index.js';
 
-const stylelint = require('../../lib');
-const { caseConfigFile, caseFiles, prepForSnapshot } = require('../systemTestUtils');
+import { caseConfigFile, caseFiles, prepForSnapshot } from '../systemTestUtils.mjs';
 
 const CASE_NUMBER = '002';
